+++ src/feedback/components/FeedbackForm.vue
@@ -10,7 +10,7 @@
     />
 
     <AmountPicker
-      v-if="!isBikeKitchen"
+      v-if="!isBikeKitchen && !isGeneralPurpose"
       v-model="edit.weight"
       class="q-ml-sm"
       style="margin-top: 40px"
@@ -63,6 +63,10 @@ export default {
   mixins: [statusMixin, editMixin],
   props: {
     isBikeKitchen: {
+      type: Boolean,
+      default: false,
+    },
+    isGeneralPurpose: {
       type: Boolean,
       default: false,
     },