+++ src/feedback/components/PickupFeedback.vue
@@ -55,6 +55,7 @@
           :value="feedbackDefault"
           :status="saveStatus"
           :is-bike-kitchen="isBikeKitchen"
+          :is-general-purpose="isGeneralPurpose"
           @save="$emit('save', arguments[0])"
         />
       </div>
@@ -144,6 +145,10 @@ export default {
       type: Number,
     },
     isBikeKitchen: {
+      type: Boolean,
+      default: false,
+    },
+    isGeneralPurpose: {
       type: Boolean,
       default: false,
     },