+++ CHANGELOG.md
@@ -22,6 +22,10 @@ Please document your changes in this format:
 
 ## [Unreleased]
 
+## [7.4.2] - 2019-08-14
+### Added
+- General purpose group theme [#1691] @djahnie
+
 ## [7.4.1] - 2019-08-09
 ### Added
 - "Unsubscribe all" button on settings page @tiltec @djahnie
@@ -472,8 +476,10 @@ After months of planning and programming, the first version of ~the foodsaving t
 [#1398]: https://github.com/yunity/karrot-frontend/issues/1398
 [#1421]: https://github.com/yunity/karrot-frontend/issues/1421
 [#1509]: https://github.com/yunity/karrot-frontend/issues/1509
+[#1691]: https://github.com/yunity/karrot-frontend/issues/1691
 
-[Unreleased]: https://github.com/yunity/karrot-frontend/compare/v7.4.1...HEAD
+[Unreleased]: https://github.com/yunity/karrot-frontend/compare/v7.4.2...HEAD
+[7.4.2]: https://github.com/yunity/karrot-frontend/compare/v7.4.1...v7.4.2
 [7.4.1]: https://github.com/yunity/karrot-frontend/compare/v7.4.0...v7.4.1
 [7.4.0]: https://github.com/yunity/karrot-frontend/compare/v7.3.0...v7.4.0
 [7.3.0]: https://github.com/yunity/karrot-frontend/compare/v7.2.4...v7.3.0